move cliff generation to cliff mod
