+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub mod cliffs;
 pub mod merge;
 pub mod util;
